fix: enusre sheet is animated correctly with new framer motion
With the framer-motion 11.0.18 update the sheet opening animation stopped working.
This commit fixes that by moving the motion animation to wrap
the dialog content and overlay.
The sheet exit animation has been broken prior to the 11.0.18 patch and this commit
also fixes that by conditionally rendering the dialog root with isOpen,
as is described in the framer motion docs.