fix: ensure alerts are animated correctly with new framer-motion
This commit is basically a copy of the changes in
d071be05e41ff407abe5ad7ccafdcaf4fa80f371 and
d0bcf4dccf652c24602786fc178210b84f003982, but in Alert.tsx.

Prior to this commit alerts would be rendered with opacity 0, making
them completely invisible to a user. These tests still worked because
the elements were in the DOM, but they did render as visible elements.

With these changes, alerts render visibly again.